+++ src/lib.rs
@@ -183,7 +183,7 @@ pub use commands::HeaderOverrides;
 #[doc(hidden)]
 pub use commands::{
     altw, apply_changes, cat, diff, getid, getparents, inspect,
-    merge_changes, renumber, sort,
+    merge_changes, renumber, repack, sort,
     tags_count, tags_filter, time_filter,
 };
 #[cfg(feature = "commands")]